test: rewrite request handler tests for react-router
